move and enhance tests
